Started styling first welcome page
Added folder containing fonts, folder containing pages (views), resources dictionaries (fonts, colors, styles).
