+++ BlazeCart/BlazeCart/App.xaml.cs
@@ -1,4 +1,6 @@
 ﻿namespace BlazeCart;
+using BlazeCart.Views;
+using BlazeCart.Resources;
 
 public partial class App : Application
 {
@@ -6,6 +8,7 @@ public partial class App : Application
 	{
 		InitializeComponent();
 
-		MainPage = new AppShell();
+		//MainPage = new AppShell();
+		MainPage = new NavigationPage(new WelcomePage1());
 	}
 }
