+++ BlazeCart/BlazeCart/Resources/Styles/Colors.xaml.cs
@@ -1,0 +1,9 @@
+namespace BlazeCart.Resources.Styles;
+
+public partial class Colors : ResourceDictionary
+{
+	public Colors()
+	{
+		InitializeComponent();
+	}
+}
