+++ BlazeCart/BlazeCart/Resources/Styles/Fonts.xaml
@@ -1,0 +1,14 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<ResourceDictionary xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             x:Class="BlazeCart.Resources.Styles.Fonts">
+
+    <OnPlatform x:Key="WelcomeFontPrimary" x:TypeArguments="x:Double">
+        <On Platform="iOS" Value="24" />
+        <On Platform="Android" Value="24" />
+    </OnPlatform>
+    <OnPlatform x:Key="WelcomeFontSecondary" x:TypeArguments="x:Double">
+        <On Platform="iOS" Value="14" />
+        <On Platform="Android" Value="14" />
+    </OnPlatform>
+</ResourceDictionary>
