+++ BlazeCart/BlazeCart/Resources/Styles/Fonts.xaml.cs
@@ -1,0 +1,9 @@
+namespace BlazeCart.Resources.Styles;
+
+public partial class Fonts : ResourceDictionary
+{
+	public Fonts()
+	{
+		InitializeComponent();
+	}
+}
