+++ BlazeCart/BlazeCart/Resources/Styles/Styles.xaml.cs
@@ -1,0 +1,9 @@
+namespace BlazeCart.Resources.Styles;
+
+public partial class Styles : ResourceDictionary
+{
+	public Styles()
+	{
+		InitializeComponent();
+	}
+}
