+++ BlazeCart/BlazeCart/Views/CartPage.xaml.cs
@@ -1,0 +1,9 @@
+namespace BlazeCart.Views;
+
+public partial class CartPage : ContentPage
+{
+	public CartPage()
+	{
+		InitializeComponent();
+	}
+}
