+++ BlazeCart/BlazeCart/Views/ErrorPage.xaml.cs
@@ -1,0 +1,9 @@
+namespace BlazeCart.Views;
+
+public partial class ErrorPage : ContentPage
+{
+	public ErrorPage()
+	{
+		InitializeComponent();
+	}
+}
