+++ BlazeCart/BlazeCart/Views/ItemCatalogPage.xaml.cs
@@ -1,0 +1,9 @@
+namespace BlazeCart.Views;
+
+public partial class ItemCatalogPage : ContentPage
+{
+	public ItemCatalogPage()
+	{
+		InitializeComponent();
+	}
+}
