+++ BlazeCart/BlazeCart/Views/WelcomePage1.xaml.cs
@@ -1,0 +1,9 @@
+namespace BlazeCart.Views;
+
+public partial class WelcomePage1 : ContentPage
+{
+	public WelcomePage1()
+	{
+		InitializeComponent();
+	}
+}
