+++ BlazeCart/BlazeCart/Views/WelcomePage2.xaml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             x:Class="BlazeCart.Views.WelcomePage2"
+             NavigationPage.HasNavigationBar="False"
+             Title="WelcomePage2">
+                
+</ContentPage>
