+++ BlazeCart/BlazeCart/Views/WelcomePage2.xaml.cs
@@ -1,0 +1,9 @@
+namespace BlazeCart.Views;
+
+public partial class WelcomePage2 : ContentPage
+{
+	public WelcomePage2()
+	{
+		InitializeComponent();
+	}
+}
