--- BlazeCart/BlazeCart/WelcomePage1.xaml.cs
@@ -1,20 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-using Xamarin.Forms;
-using Xamarin.Forms.Xaml;
-
-namespace BlazeCart
-{
-    [XamlCompilation(XamlCompilationOptions.Compile)]
-    public partial class Page1 : ContentPage
-    {
-        public Page1()
-        {
-            InitializeComponent();
-        }
-    }
-}
